chat/pyproject.toml
[project]
name = "Eirene"
version = "0.1.0"
description = "API Server for Multi-Agent AI Care System for End-of-Life Patients."
readme = "README.md"
requires-python = ">=3.12.0, <3.13"
dependencies = [
    "backoff>=2.2.1",
    "bs4>=0.0.2",
    "faiss-cpu>=1.11.0",
    "fastapi[standard]>=0.116.1",
    "konlpy>=0.6.0",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-core>=0.3.66",
    "langchain-naver>=0.1.0",
    "langchain-openai>=0.3.25",
    "langgraph>=0.5.0",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "posthog>=6.1.0",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.0",
    "pytz>=2025.2",
    "scikit-learn>=1.7.1",
    "uvicorn>=0.35.0",
]
